requirements.txt
aiogram
sqlalchemy
python-dotenv
beautifulsoup4
requests
pymystem3
loguru